% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{computePCEV}
\alias{computePCEV}
\title{Principal Component of Explained Variance}
\usage{
computePCEV(response, covariate, confounder = NULL, estimation = "all",
  inference = "exact", index = NULL, shrink = FALSE, nperm = 1000,
  Wilks = FALSE, reduce = TRUE)
}
\arguments{
\item{response}{A matrix of response variables.}

\item{covariate}{A matrix or a data frame of covariates.}

\item{confounder}{A matrix or data frame of confounders}

\item{estimation}{Character string specifying which estimation method to use:
\code{"all"} or \code{"block"}. Default value is \code{"all"}.}

\item{inference}{Character string specifying which inference method to use:
\code{"exact"} or \code{"permutation"}. Default value is \code{"exact"}.}

\item{index}{If \code{estimation = "block"}, index is a vector describing the
block to which individual response variables correspond.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?
Default value is \code{FALSE}..}

\item{nperm}{The number of permutations to perform if \code{inference =
"permutation"}}

\item{Wilks}{Should we use a Wilks test instead of Roy's largest test? This
is only implemented for a single covariates.}

\item{reduce}{Should we use the previous estimates of the residual variance matrix?
See details. The default value is \code{TRUE}.}
}
\value{
An object of class \code{Pcev} containing the first PCEV, the
  p-value, the estimate of the shrinkage factor, etc.
}
\description{
\code{computePCEV} computes the first PCEV and tests its significance.
}
\details{
This is the main function. It computes the PCEV using either the classical
method or the block approach. A p-value is also computed, testing the
significance of the PCEV.

When a permutation procedure is used to either estimate a p-value or estimate
the parameters of the null distribution under shrinkage, computations can be
considerably sped up by re-using previous estimates of the residual variance
matrix. This is the default behaviour, and it can be changed by setting the
parameter \code{reduce} to \code{FALSE}.
}
\examples{
set.seed(12345)
Y <- matrix(rnorm(100*20), nrow=100)
X <- rnorm(100)
pcev_out <- computePCEV(Y, X)
pcev_out2 <- computePCEV(Y, X, shrink = TRUE)
}
\seealso{
\code{\link{estimatePcev}}
}

