% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimatePcev_red.R
\name{estimatePcev_red}
\alias{estimatePcev_red}
\alias{estimatePcev_red.PcevBlock}
\alias{estimatePcev_red.PcevClassical}
\alias{estimatePcev_red.default}
\title{Estimation of PCEV (reduced computation)}
\usage{
estimatePcev_red(pcevObj, ...)

\method{estimatePcev_red}{default}(pcevObj, ...)

\method{estimatePcev_red}{PcevClassical}(pcevObj, index, root_Vr, ...)

\method{estimatePcev_red}{PcevBlock}(pcevObj, index, root_Vr_list, ...)
}
\arguments{
\item{pcevObj}{A pcev object of class \code{PcevClassical} or
\code{PcevBlock}}

\item{...}{Extra parameters.}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, index is a
vector describing the block to which individual response variables
correspond.}

\item{root_Vr}{Previous estimate of the square root of the inverse of the
residual variance}

\item{root_Vr_list}{List of previous estimates of the square root of the
inverse of the residual variances. Requires one for each block and one for
the second stage of the estimation.}
}
\value{
A list containing the first PCEV and the largest eigenvalue of
  \eqn{V_R^{-1}V_G}.
}
\description{
\code{estimatePcev_red} estimates the PCEV by reusing previous estimate of
the residual variance. This is used to speed up computations of the
permutation procedures.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{PcevClassical}: 

\item \code{PcevBlock}: 
}}
\seealso{
\code{\link{computePCEV}}
}

