% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{influence.Pcev}
\alias{influence.Pcev}
\title{Influence measure on the PCEV}
\usage{
\method{influence}{Pcev}(model, ...)
}
\arguments{
\item{model}{an object of class \code{Pcev}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Influence measures
}
\description{
This function computes a measure of the influence of each observation on the PCEV.
}
\details{
This measure is computed following equation 14 in Pendergast (2011).
}

