% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{permutePval}
\alias{permutePval}
\alias{permutePval.PcevBlock}
\alias{permutePval.PcevClassical}
\alias{permutePval.default}
\title{Permutation p-value}
\usage{
permutePval(pcevObj, ...)

\method{permutePval}{default}(pcevObj, ...)

\method{permutePval}{PcevClassical}(pcevObj, shrink, index, nperm, ...)

\method{permutePval}{PcevBlock}(pcevObj, shrink, index, nperm, ...)
}
\arguments{
\item{pcevObj}{A pcev object of class \code{PcevClassical} or
\code{PcevBlock}}

\item{...}{Extra parameters.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, index is a
vector describing the block to which individual response variables
correspond.}

\item{nperm}{The number of permutations to perform.}
}
\description{
Computes a p-value using a permutation procedure.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{PcevClassical}: 

\item \code{PcevBlock}: 
}}

