% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roysPval}
\alias{roysPval}
\alias{roysPval.PcevBlock}
\alias{roysPval.PcevClassical}
\alias{roysPval.PcevSingular}
\alias{roysPval.default}
\title{Roy's largest root exact test}
\usage{
roysPval(pcevObj, ...)

\method{roysPval}{default}(pcevObj, ...)

\method{roysPval}{PcevClassical}(pcevObj, shrink, index, ...)

\method{roysPval}{PcevSingular}(pcevObj, shrink, distrib, index, ...)

\method{roysPval}{PcevBlock}(pcevObj, shrink, index, ...)
}
\arguments{
\item{pcevObj}{A pcev object of class \code{PcevClassical} or 
\code{PcevBlock}}

\item{...}{Extra parameters.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, \code{index} is
a vector describing the block to which individual response variables 
correspond}

\item{distrib}{If \code{estimation = "singular"}, choose one of two statistics,
\code{"Wishart"} or \code{"TW"}.}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, \code{index} is a 
vector describing the block to which individual response variables}

\item{method}{Method to use for the estimation of the location-scale
parameters. The available methods are the method of moments (default) and
Maximum Likelihood Estimation.}
}
\description{
In the classical domain of PCEV applicability this function uses Johnstone's
approximation to the null distribution of ' Roy's Largest Root statistic.
It uses a location-scale variant of the Tracy-Wildom distribution of order 1.
}
\details{
For singular PCEV, where number of variables is higher than the number of observations,
one can choose between two aprroximations. First one, called 'Wishart', is based
on the first term of expansion of the joint distribution of roots for singular beta
ensemble. Second version of test, \code{estimation = "TW"}, assumes that largest root statistics follows
Tracy-Wildom as in classical case. See vignette for the domains of the applicability
of these tests.

Note that if \code{shrink} is set to \code{TRUE}, the location-scale 
parameters are estimated using a small number of permutations.
}

