% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{WilksLambda}
\alias{Wilks}
\alias{WilksLambda}
\alias{Wilkslambda}
\alias{wilks}
\alias{wilksLambda}
\alias{wilkslambda}
\title{Wilks Lambda test.}
\usage{
WilksLambda(Y, x, shrink = FALSE)
}
\arguments{
\item{Y}{Matrix of values for a multivariate response.}

\item{x}{Vector of values for a covariate.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?
Default value is FALSE.}
}
\value{
A list containing the variance components, the first PCEV, the
  p-value, etc.
}
\description{
\code{WilksLambda} computes a p-value for the classical PCEV.
}
\details{
The Wilks Lambda essentially tests whether the proportion of variance being
explained by the covariates is zero. This function computes the variance
decomposition which is used to compute the first PCEV. Note that this method
is currently only implemented for the case where there is only one covariate.
}
\seealso{
\code{\link{computePCEV}}
}

