% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{wilksPval}
\alias{wilksPval}
\alias{wilksPval.PcevBlock}
\alias{wilksPval.PcevClassical}
\alias{wilksPval.default}
\title{Wilks' lambda exact test}
\usage{
wilksPval(pcevObj, ...)

\method{wilksPval}{default}(pcevObj, ...)

\method{wilksPval}{PcevClassical}(pcevObj, shrink, index)

\method{wilksPval}{PcevBlock}(pcevObj, shrink, index)
}
\arguments{
\item{pcevObj}{A pcev object of class \code{PcevClassical} or
\code{PcevBlock}}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, index is a vector
describing the block to which individual response variables correspond.}
}
\description{
sdf
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{PcevClassical}: 

\item \code{PcevBlock}: 
}}

