% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{methylation}
\alias{index}
\alias{methylation}
\alias{pheno}
\alias{position}
\title{Methylation values around BLK gene}
\format{The data comes in four objects:
\describe{
\item{methylation}{Matrix of methylation values at 5,986 sites measured on 40 samples}
\item{pheno}{Vector of phenotype, indicating whether the sample comes from a B cell}
\item{position}{Data frame recording the position of each CpG site along the BLK region}
\item{index}{Index vector used in the computation of PCEV-block}
}}
\source{
Tomi Pastinen, McGill University
}
\usage{
methylation

pheno

position

index
}
\description{
A dataset containing methylation values for cell-separated samples. The
methylation was measured using bisulfite sequencing. The data also contains
the genomic position of these CpG sites, as well as a binary phenotype
(i.e. whether the sample comes from a B cell).
}
\details{
Methylation was first measured at 24,068 sites, on 40 samples. Filtering was
performed to keep the 25\% most variable sites. See the vignette for more detail.
}
\keyword{datasets}

