% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{wilksPval}
\alias{wilksPval}
\alias{wilksPval.default}
\alias{wilksPval.PcevClassical}
\alias{wilksPval.PcevSingular}
\alias{wilksPval.PcevBlock}
\title{Wilks' lambda exact test}
\usage{
wilksPval(pcevObj, ...)

\method{wilksPval}{default}(pcevObj, ...)

\method{wilksPval}{PcevClassical}(pcevObj, shrink, index, ...)

\method{wilksPval}{PcevSingular}(pcevObj, shrink, index, ...)

\method{wilksPval}{PcevBlock}(pcevObj, shrink, index, ...)
}
\arguments{
\item{pcevObj}{A pcev object of class \code{PcevClassical} or 
\code{PcevBlock}}

\item{...}{Extra parameters.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, \code{index} is
a vector describing the block to which individual response variables 
correspond.}
}
\description{
Computes a p-value using Wilks' Lambda.
}
\details{
The null distribution of this test statistic is only known in the case of a
single covariate, and therefore this is the only case implemented.
}
